requirements.txt
qdrant-client
markdown
transformers
langchain
langchain-community
langchain-huggingface
langchain-qdrant
beautifulsoup4
python-dotenv